read text, extract place names, plot points
This is the completed version of the program that reads the text of Fawcett's explorations, extracts candidate place names and uses a geocoding service to get lattitudes and longitudes for the place names, and finally plots these as points on a map of South America. 
Its taken me a while, it is by no means perfect, but it will serve as a basis to learn more about how to to obtain and access data about geographic locations.